move near into protocols subfolder
gonna try keeping this around for now, and put a `cw` or `cosmwasm`
folder beside it.

Not sure if I'll leave both in this version of the app, but it will be
nice to reference for now.
